+++ src/screens/appointments/Attended.js
@@ -1,0 +1,17 @@
+import React, { Component } from "react";
+import { Text, View } from "react-native";
+
+export default class Attended extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
+
+  render() {
+    return (
+      <View>
+        <Text> TAB Attended </Text>
+      </View>
+    );
+  }
+}
